benchmark/php-mysql/requirements.txt
-e .
pytest
aws-cdk.aws-ec2
aws-cdk.aws-rds
aws-cdk.core
aws-cdk.iam
